chore: add home page
